pset/SG-I2V/requirements.txt
accelerate==0.31.0
argparse
diffusers[torch]==0.29.1
imageio==2.34.1
ipython==8.12.3
matplotlib==3.9.2
numpy==2.1.3
opencv_python==4.10.0.84
Pillow==11.0.0
pillow_avif_plugin==1.4.6
pytorch_lightning==2.3.0
torch==2.3.1
torchvision==0.18.1
transformers==4.41.2
